State - Take a ref on an object for each state
The state is linked into the list in the state_hdl.  For MDCACHE, this
lives directly in the entry itself.  This means that, if the entry is
reused, this list get's cleaned and reused.  The new entry can then be
freed, and when the old state is removed, it derefences into the now
freed entry, causing use-after-free.

Take a ref on the entry for each state; this will keep the entry from
being reused off the LRU.

Change-Id: I8d810c2cb884450c08620096f77d08fe70255d8d
Signed-off-by: Daniel Gryniewicz <dang@redhat.com>